pyproject.toml
[project]
name = "hr-breaker"
version = "0.1.0"
description = "Resume optimization tool for job postings"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Boris Tseitlin"}]
license = {text = "MIT"}
keywords = ["resume", "job", "optimization", "ats", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "streamlit>=1.30",
    "pydantic>=2.0",
    "pydantic-ai",
    "openai",
    "httpx",
    "beautifulsoup4",
    "spacy>=3.0",
    "qdrant-client",
    "sentence-transformers",
    "scikit-learn>=1.0",
    "python-dotenv",
    "click>=8.0",
    "pymupdf>=1.24",
    "importlib-metadata>=6.0,<8.0",
    "weasyprint>=60.0",
    "jinja2>=3.1",
    "watchdog>=6.0.0",
    "nest-asyncio>=1.6.0",
    "playwright>=1.40",
]

[project.scripts]
hr-breaker = "hr_breaker.cli:cli"

[project.urls]
Homepage = "https://github.com/btseytlin/hr-breaker"
Repository = "https://github.com/btseytlin/hr-breaker"
Issues = "https://github.com/btseytlin/hr-breaker/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hr_breaker"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
